# 8356593: RISC-V: Small improvement to array fill stub