When working on [JDK-8351140](https://bugs.openjdk.org/browse/JDK-8351140), I witnessed possible misaligned memory access in array fill stub.
We fill by element for short arrays (< 8 bytes), which assumes a heapword alignment[1]. But that is not guaranteed. This issue could be reproduced by running: `make test TEST="micro:vm.compiler.ArrayFill"`
with `@Param("5") private int size;` on riscv platforms with slow misalgned memory accesses.

[1] https://github.com/openjdk/jdk/blob/master/src/hotspot/cpu/riscv/stubGenerator_riscv.cpp#L2141

This fixes this issue by using a small loop to fill the elements for short arrays.
Tier1-2 tested on linux-riscv64 platform. JMH result on P550 SBC for reference:
1. (`@Param("5") private int size;`):
```
Before:
Benchmark                 (size)  Mode  Cnt    Score   Error  Units
ArrayFill.fillByteArray        5  avgt   12  558.781 ± 1.396  ns/op
ArrayFill.fillIntArray         5  avgt   12   29.346 ± 0.003  ns/op
ArrayFill.fillShortArray       5  avgt   12   30.779 ± 0.004  ns/op
ArrayFill.zeroByteArray        5  avgt   12  559.249 ± 1.909  ns/op
ArrayFill.zeroIntArray         5  avgt   12   29.346 ± 0.002  ns/op
ArrayFill.zeroShortArray       5  avgt   12   30.777 ± 0.006  ns/op

After:
Benchmark                 (size)  Mode  Cnt   Score   Error  Units
ArrayFill.fillByteArray        5  avgt   12  23.977 ± 0.004  ns/op
ArrayFill.fillIntArray         5  avgt   12  29.343 ± 0.004  ns/op
ArrayFill.fillShortArray       5  avgt   12  30.776 ± 0.005  ns/op
ArrayFill.zeroByteArray        5  avgt   12  23.977 ± 0.002  ns/op
ArrayFill.zeroIntArray         5  avgt   12  29.345 ± 0.005  ns/op
ArrayFill.zeroShortArray       5  avgt   12  30.776 ± 0.004  ns/op
```
2. (`@Param("3") private int size;`):
```
Before:
Benchmark                 (size)  Mode  Cnt    Score   Error  Units
ArrayFill.fillByteArray        3  avgt   12  428.923 ± 0.409  ns/op
ArrayFill.fillIntArray         3  avgt   12   28.629 ± 0.005  ns/op
ArrayFill.fillShortArray       3  avgt   12  558.872 ± 2.641  ns/op
ArrayFill.zeroByteArray        3  avgt   12  429.744 ± 2.049  ns/op
ArrayFill.zeroIntArray         3  avgt   12   28.628 ± 0.002  ns/op
ArrayFill.zeroShortArray       3  avgt   12  557.682 ± 1.661  ns/op

After:
Benchmark                 (size)  Mode  Cnt   Score   Error  Units
ArrayFill.fillByteArray        3  avgt   12  21.471 ± 0.002  ns/op
ArrayFill.fillIntArray         3  avgt   12  28.631 ± 0.003  ns/op
ArrayFill.fillShortArray       3  avgt   12  20.436 ± 0.288  ns/op
ArrayFill.zeroByteArray        3  avgt   12  21.472 ± 0.002  ns/op
ArrayFill.zeroIntArray         3  avgt   12  28.633 ± 0.003  ns/op
ArrayFill.zeroShortArray       3  avgt   12  20.203 ± 0.137  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8356593](https://bugs.openjdk.org/browse/JDK-8356593): RISC-V: Small improvement to array fill stub (**Enhancement** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/25135/head:pull/25135` \
`$ git checkout pull/25135`

Update a local copy of the PR: \
`$ git checkout pull/25135` \
`$ git pull https://git.openjdk.org/jdk.git pull/25135/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 25135`

View PR using the GUI difftool: \
`$ git pr show -t 25135`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/25135.diff">https://git.openjdk.org/jdk/pull/25135.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/25135#issuecomment-2865417332)
</details>