@Anjian-Wen 2025-05-09
/issue JDK-8356593