@RealFYang 2025-05-10
PS: Would you mind adding following small change while you are on it? This optimizes out three register-register moves.
[25135-addon.diff.txt](https://github.com/user-attachments/files/20133615/25135-addon.diff.txt)