@Anjian-Wen 2025-05-10
Thats looks good, done!

> PS: Would you mind adding following small change while you are on it? This optimizes out three register-register moves in this array fill stub. [25135-addon.diff.txt](https://github.com/user-attachments/files/20133615/25135-addon.diff.txt)